% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{acor}
\alias{acor}
\title{computing the autocorrelation of time series data}
\usage{
acor(y, acor_order)
}
\arguments{
\item{y}{vector of time series data}

\item{acor_order}{positive integer for the order of the autocorrelation}
}
\description{
computing the autocorrelation of time series data
}
