% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{acov}
\alias{acov}
\title{computing the autocovariance of time series data}
\usage{
acov(y, acov_order)
}
\arguments{
\item{y}{vector of time series data}

\item{acov_order}{non-negative integer for the order of the autocovariance}
}
\description{
computing the autocovariance of time series data
}
