% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{ecdfest}
\alias{ecdfest}
\title{computing empirical CDF estimate}
\usage{
ecdfest(x, X)
}
\arguments{
\item{x}{point at which the CDF is estimated}

\item{X}{vector of cross-sectional data}
}
\description{
computing empirical CDF estimate
}
