% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error1}
\alias{error1}
\title{function 1 to handle errors}
\usage{
error1(data, acov_order, acor_order, R)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}

\item{R}{positive integer for the number of bootstrap replications}
}
\description{
function 1 to handle errors
}
