% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error2}
\alias{error2}
\title{function 2 to handle errors}
\usage{
error2(data, acov_order, acor_order)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}
}
\description{
function 2 to handle errors
}
