% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error3}
\alias{error3}
\title{function 3 to handle errors}
\usage{
error3(data, acov_order, acor_order, mean_bw, acov_bw, acor_bw)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}

\item{mean_bw}{bandwidth for the mean}

\item{acov_bw}{bandwidth for the autocovariance}

\item{acor_bw}{bandwidth for the autocorrelation}
}
\description{
function 3 to handle errors
}
