% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfhpj.R
\name{hpj1_boot}
\alias{hpj1_boot}
\title{Compute bootstrap HPJ empirical CDF estimate for odd T}
\usage{
hpj1_boot(quantity, indices)
}
\arguments{
\item{quantity}{An N * 3 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap HPJ empirical CDF estimate for odd T
}
