% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfhpj.R
\name{hpj2_boot}
\alias{hpj2_boot}
\title{computing bootstrap HPJ empirical CDF estimate for even T}
\usage{
hpj2_boot(quantity, indices)
}
\arguments{
\item{quantity}{N * 5 matrix of estimates}

\item{indices}{indices for bootstrap replications}
}
\description{
computing bootstrap HPJ empirical CDF estimate for even T
}
