% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfhpj.R
\name{hpjecdf}
\alias{hpjecdf}
\title{HPJ bias-corrected empirical CDF estimation for heterogeneity in panel data}
\usage{
hpjecdf(data, acov_order = 0, acor_order = 1)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}
}
\value{
list that contains the following elements.
\item{mean}{graph of the HPJ bias-corrected empirical CDF estimation for the mean with rearrangement}
\item{acov}{graph of the HPJ bias-corrected empirical CDF estimation for the autocovariance with rearrangement}
\item{acor}{graph of the HPJ bias-corrected empirical CDF estimation for the autocorrelation with rearrangement}
\item{mean_func}{function that returns HPJ bias-corrected empirical CDF estimates for the mean without rearrangement}
\item{acov_func}{function that returns HPJ bias-corrected empirical CDF estimates for the autocovariance without rearrangement}
\item{acor_func}{function that returns HPJ bias-corrected empirical CDF estimates for the autocorrelation without rearrangement}
\item{quantity}{matrix that contains the estimated quantities}
\item{acov_order}{the same as the argument}
\item{acor_order}{the same as the argument}
\item{N}{the number of cross-sectional units}
\item{S}{the length of time series}
}
\description{
\code{hpjecdf} implements the HPJ bias-corrected estimation of
the empirical CDF for the heterogeneous mean, autocovariance,
and autocorrelation.
The procedure is proposed in Okui and Yanagi (2017).
See the package vignette via \code{vignette("panelhetero")} for details.
}
