% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfhpj.R
\name{hpjecdfest1}
\alias{hpjecdfest1}
\title{Compute HPJ bias-corrected empirical CDF estimate for even T}
\usage{
hpjecdfest1(x, X, X1, X2)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of original cross-sectional data}

\item{X1}{A vector of half-panel cross-sectional data 1}

\item{X2}{A vector of half-panel cross-sectional data 2}
}
\description{
Compute HPJ bias-corrected empirical CDF estimate for even T
}
