% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfhpj.R
\name{hpjecdfest2}
\alias{hpjecdfest2}
\title{computing HPJ empirical CDF estimate for odd T}
\usage{
hpjecdfest2(x, X, X1, X2, X3, X4)
}
\arguments{
\item{x}{point at which the CDF is estimated}

\item{X}{vector of original cross-sectional data}

\item{X1}{vector of half-panel cross-sectional data based on time series 1 ~ floor(T/2)}

\item{X2}{vector of half-panel cross-sectional data based on time series (floor(T/2) + 1) ~ T}

\item{X3}{vector of half-panel cross-sectional data based on time series 1 ~ ceiling(T/2)}

\item{X4}{vector of half-panel cross-sectional data based on time series (ceiling(T/2) + 1) ~ T}
}
\description{
computing HPJ empirical CDF estimate for odd T
}
