% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityhpj.R
\name{hpjkd}
\alias{hpjkd}
\title{HPJ bias-corrected kernel density estimation for heterogeneity in panel data}
\usage{
hpjkd(data, acov_order = 0, acor_order = 1, mean_bw = NULL,
  acov_bw = NULL, acor_bw = NULL)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}

\item{mean_bw}{bandwidth for the mean}

\item{acov_bw}{bandwidth for the autocovariance}

\item{acor_bw}{bandwidth for the autocorrelation}
}
\value{
list that contains the following elements.
\item{mean}{graph of the HPJ bias-corrected kernel density estimation for the mean}
\item{acov}{graph of the HPJ bias-corrected kernel density estimation for the autocovariance}
\item{acor}{graph of the HPJ bias-corrected kernel density estimation for the autocorrelation}
\item{mean_func}{function that returns HPJ bias-corrected kernel density estimates for the mean}
\item{acov_func}{function that returns HPJ bias-corrected kernel density estimates for the autocovariance}
\item{acor_func}{function that returns HPJ bias-corrected kernel density estimates for the autocorrelation}
\item{bandwidth}{vector of the selected bandwidths}
\item{quantity}{matrix that contains the estimated quantities}
\item{acov_order}{the same as the argument}
\item{acor_order}{the same as the argument}
\item{N}{the number of cross-sectional units}
\item{S}{the length of time series}
}
\description{
\code{hpjkd} implements the HPJ bias-corrected estimation of
the kernel density for the heterogeneous mean, autocovariance,
and autocorrelation.
The procedure is proposed in Okui and Yanagi (2018).
See the package vignette via \code{vignette("panelhetero")} for details.
}
