% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityhpj.R
\name{hpjkdest1}
\alias{hpjkdest1}
\title{computing HPJ kernel density estimate for even T}
\usage{
hpjkdest1(x, X, X1, X2, h)
}
\arguments{
\item{x}{point at which the density is estimated}

\item{X}{vector of original cross-sectional data}

\item{X1}{vector of half-panel cross-sectional data based on time series 1 ~ T/2}

\item{X2}{vector of half-panel cross-sectional data based on time series (T/2 + 1) ~ T}

\item{h}{bandwidth}
}
\description{
computing HPJ kernel density estimate for even T
}
