% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityhpj.R
\name{hpjkdest2}
\alias{hpjkdest2}
\title{computing HPJ kernel density estimate for odd T}
\usage{
hpjkdest2(x, X, X1, X2, X3, X4, h)
}
\arguments{
\item{x}{point at which the density is estimated}

\item{X}{vector of original cross-sectional data}

\item{X1}{vector of half-panel cross-sectional data based on time series 1 ~ floor(T/2)}

\item{X2}{vector of half-panel cross-sectional data based on time series (floor(T/2) + 1) ~ T}

\item{X3}{vector of half-panel cross-sectional data based on time series 1 ~ ceiling(T/2)}

\item{X4}{vector of half-panel cross-sectional data based on time series (ceiling(T/2) + 1) ~ T}

\item{h}{bandwidth}
}
\description{
computing HPJ kernel density estimate for odd T
}
