% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenthpj.R
\name{hpjmomentest1}
\alias{hpjmomentest1}
\title{computing HPJ bias-corrected estimates of the moments for even T}
\usage{
hpjmomentest1(quantity, indices)
}
\arguments{
\item{quantity}{N * 3 matrix of the means, autocovariances, autocorrelations}

\item{indices}{indices for bootstrap replications}
}
\description{
computing HPJ bias-corrected estimates of the moments for even T
}
