% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenthpj.R
\name{hpjmomentest2}
\alias{hpjmomentest2}
\title{computing HPJ bias-corrected estimates of the moments for odd T}
\usage{
hpjmomentest2(quantity, indices)
}
\arguments{
\item{quantity}{N * 3 matrix of the estimated means, autocovariances, autocorrelations}

\item{indices}{indices for bootstrap replications}
}
\description{
computing HPJ bias-corrected estimates of the moments for odd T
}
