% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcdf.R
\name{kcdfest}
\alias{kcdfest}
\title{computing kernel CDF estimate}
\usage{
kcdfest(x, X, h)
}
\arguments{
\item{x}{point at which the CDF is estimated}

\item{X}{vector of cross-sectional data}

\item{h}{bandwidth}
}
\description{
computing kernel CDF estimate
}
