% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{kdest}
\alias{kdest}
\title{computing kernel density estimate}
\usage{
kdest(x, X, h)
}
\arguments{
\item{x}{point at which the density is estimated}

\item{X}{vector of cross-sectional data}

\item{h}{bandwidth}
}
\description{
computing kernel density estimate
}
