% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{momentest}
\alias{momentest}
\title{computing estimates for the moments}
\usage{
momentest(quantity, indices)
}
\arguments{
\item{quantity}{N * 3 matrix of the means, autocovariances, autocorrelations}

\item{indices}{indices for bootstrap replications}
}
\description{
computing estimates for the moments
}
