% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{ne_boot}
\alias{ne_boot}
\title{computing bootstrap empirical CDF estimate}
\usage{
ne_boot(quantity, indices)
}
\arguments{
\item{quantity}{N * 1 matrix of estimate}

\item{indices}{indices for bootstrap replications}
}
\description{
computing bootstrap empirical CDF estimate
}
