% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{nemoment}
\alias{nemoment}
\title{Estimation of moments for heterogeneity in panel data without bias-correction}
\usage{
nemoment(data, acov_order = 0, acor_order = 1, R = 1000)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}

\item{R}{positive integer for the number of bootstrap replications}
}
\value{
list that contains the following elements.
\item{estimate}{vector of estimates of the parameters}
\item{se}{vector of standard errors of the estimators}
\item{ci}{matrix of 95 percent confidence intervals of the parameters}
\item{quantity}{matrix that contains the estimated quantities}
\item{acov_order}{the same as the argument}
\item{acor_order}{the same as the argument}
\item{N}{the number of cross-sectional units}
\item{S}{the length of time series}
\item{R}{the number of bootstrap replications}
}
\description{
\code{nemoment} implements the naive estimation of the moments
for the heterogeneous mean, autocovariance, and autocorrelation.
The procedure is proposed in Okui and Yanagi (2017).
See the package vignette via \code{vignette("panelhetero")} for details.
}
