% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Generate simulated data}
\usage{
simulation(N, S)
}
\arguments{
\item{N}{The number of cross-sectional units}

\item{S}{The length of time series}
}
\value{
An N times S matrix of panel data
}
\description{
Generate simulated data
}
\examples{
panelhetero:::simulation(N = 300, S = 50)

}
