% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{toj4_boot}
\alias{toj4_boot}
\title{computing bootstrap TOJ empirical CDF estimate for even T}
\usage{
toj4_boot(quantity, indices)
}
\arguments{
\item{quantity}{N * 12 matrix of estimates}

\item{indices}{indices for bootstrap replications}
}
\description{
computing bootstrap TOJ empirical CDF estimate for even T
}
