% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{tojecdf}
\alias{tojecdf}
\title{TOJ bias-corrected empirical CDF estimation for heterogeneity in panel data}
\usage{
tojecdf(data, acov_order = 0, acor_order = 1, R = 1000, ci = TRUE)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}

\item{R}{positive integer for the number of bootstrap replications}

\item{ci}{bool for the confidence interval}
}
\value{
list that contains the following elements.
\item{mean}{graph of the TOJ bias-corrected empirical CDF estimation for the mean with rearrangement}
\item{acov}{graph of the TOJ bias-corrected empirical CDF estimation for the autocovariance with rearrangement}
\item{acor}{graph of the TOJ bias-corrected empirical CDF estimation for the autocorrelation with rearrangement}
\item{mean_func}{function that returns TOJ bias-corrected empirical CDF estimates for the mean without rearrangement}
\item{acov_func}{function that returns TOJ bias-corrected empirical CDF estimates for the autocovariance without rearrangement}
\item{acor_func}{function that returns TOJ bias-corrected empirical CDF estimates for the autocorrelation without rearrangement}
\item{mean_ci_func}{function that returns 95 percent bootstrap confidence interval for TOJ empirical CDF estimates for the mean}
\item{acov_ci_func}{function that returns 95 percent bootstrap confidence interval for TOJ empirical CDF estimates for the autocovariance}
\item{acor_ci_func}{function that returns 95 percent bootstrap confidence interval for TOJ empirical CDF estimates for the autocorrelation}
\item{quantity}{matrix that contains the estimated quantities}
\item{acov_order}{the same as the argument}
\item{acor_order}{the same as the argument}
\item{N}{the number of cross-sectional units}
\item{S}{the length of time series}
\item{R}{the number of bootstrap replications}
}
\description{
\code{tojecdf} implements the TOJ bias-corrected estimation of
the empirical CDF for the heterogeneous mean, autocovariance,
and autocorrelation.
The procedure is proposed in Okui and Yanagi (2019).
See the package vignette via \code{vignette("panelhetero")} for details.
}
