% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{tojecdfest2}
\alias{tojecdfest2}
\title{Compute TOJ empirical CDF estimate for T equivalent to 2 modulo 6}
\usage{
tojecdfest2(x, X, X21, X22, X31, X32, X33, X34, X35, X36, X37, X38, X39)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of original cross-sectional data}

\item{X21}{A vector of half-panel cross-sectional data 1}

\item{X22}{A vector of half-panel cross-sectional data 2}

\item{X31}{A vector of third-panel cross-sectional data 1}

\item{X32}{A vector of third-panel cross-sectional data 2}

\item{X33}{A vector of third-panel cross-sectional data 3}

\item{X34}{A vector of third-panel cross-sectional data 4}

\item{X35}{A vector of third-panel cross-sectional data 5}

\item{X36}{A vector of third-panel cross-sectional data 6}

\item{X37}{A vector of third-panel cross-sectional data 7}

\item{X38}{A vector of third-panel cross-sectional data 8}

\item{X39}{A vector of third-panel cross-sectional data 9}
}
\description{
Compute TOJ empirical CDF estimate for T equivalent to 2 modulo 6
}
