% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{tojecdfest3}
\alias{tojecdfest3}
\title{Compute TOJ empirical CDF estimate for T equivalent to 3 modulo 6}
\usage{
tojecdfest3(x, X, X21, X22, X23, X24, X31, X32, X33)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of original cross-sectional data}

\item{X21}{A vector of half-panel cross-sectional data 1}

\item{X22}{A vector of half-panel cross-sectional data 2}

\item{X23}{A vector of half-panel cross-sectional data 3}

\item{X24}{A vector of half-panel cross-sectional data 4}

\item{X31}{A vector of third-panel cross-sectional data 1}

\item{X32}{A vector of third-panel cross-sectional data 2}

\item{X33}{A vector of third-panel cross-sectional data 3}
}
\description{
Compute TOJ empirical CDF estimate for T equivalent to 3 modulo 6
}
