% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcdftoj.R
\name{tojkcdf}
\alias{tojkcdf}
\title{TOJ bias-corrected kernel CDF estimation for heterogeneity in panel data}
\usage{
tojkcdf(
  data,
  acov_order = 0,
  acor_order = 1,
  mean_bw = NULL,
  acov_bw = NULL,
  acor_bw = NULL
)
}
\arguments{
\item{data}{matrix of panel data in which each row is individual time series}

\item{acov_order}{non-negative integer for the order of the autocovariance}

\item{acor_order}{positive integer for the order of the autocorrelation}

\item{mean_bw}{bandwidth for the mean}

\item{acov_bw}{bandwidth for the autocovariance}

\item{acor_bw}{bandwidth for the autocorrelation}
}
\value{
list that contains the following elements.
\item{mean}{graph of the TOJ bias-corrected kernel CDF estimation for the mean with rearrangement}
\item{acov}{graph of the TOJ bias-corrected kernel CDF estimation for the autocovariance with rearrangement}
\item{acor}{graph of the TOJ bias-corrected kernel CDF estimation for the autocorrelation with rearrangement}
\item{mean_func}{function that returns TOJ bias-corrected kernel CDF estimates for the mean without rearrangement}
\item{acov_func}{function that returns TOJ bias-corrected kernel CDF estimates for the autocovariance without rearrangement}
\item{acor_func}{function that returns TOJ bias-corrected kernel CDF estimates for the autocorrelation without rearrangement}
\item{bandwidth}{vector of the selected bandwidths}
\item{quantity}{matrix that contains the estimated quantities}
\item{acov_order}{the same as the argument}
\item{acor_order}{the same as the argument}
\item{N}{the number of cross-sectional units}
\item{S}{the length of time series}
}
\description{
\code{tojkcdf} implements the TOJ bias-corrected estimation of
the kernel density for the heterogeneous mean, autocovariance,
and autocorrelation.
The procedure is proposed in Okui and Yanagi (2019).
See the package vignette via \code{vignette("panelhetero")} for details.
}
