% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcdftoj.R
\name{tojkcdfest3}
\alias{tojkcdfest3}
\title{computing TOJ kernel CDF estimate for T equivalent to 3 modulo 6}
\usage{
tojkcdfest3(x, X, X21, X22, X23, X24, X31, X32, X33, h)
}
\arguments{
\item{x}{point at which the CDF is estimated}

\item{X}{vector of original cross-sectional data}

\item{X21}{vector of half-panel cross-sectional data based on time series 1 ~ floor(T/2)}

\item{X22}{vector of half-panel cross-sectional data based on time series (floor(T/2) + 1) ~ T}

\item{X23}{vector of half-panel cross-sectional data based on time series 1 ~ ceiling(T/2)}

\item{X24}{vector of half-panel cross-sectional data based on time series (ceiling(T/2) + 1) ~ T}

\item{X31}{vector of one-third-panel cross-sectional data based on time series 1 ~ T/3}

\item{X32}{vector of one-third-panel cross-sectional data based on time series (T/3 + 1) ~ 2 * T/3}

\item{X33}{vector of one-third-panel cross-sectional data based on time series 2 * T/3 + 1 ~ T}

\item{h}{bandwidth}
}
\description{
computing TOJ kernel CDF estimate for T equivalent to 3 modulo 6
}
