% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitytoj.R
\name{tojkdest1}
\alias{tojkdest1}
\title{Compute TOJ kernel density estimate for T equivalent to 1 modulo 6}
\usage{
tojkdest1(
  x,
  X,
  X21,
  X22,
  X23,
  X24,
  X31,
  X32,
  X33,
  X34,
  X35,
  X36,
  X37,
  X38,
  X39,
  h
)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of cross-sectional data}

\item{X21}{A vector of half-panel cross-sectional data 21}

\item{X22}{A vector of half-panel cross-sectional data 22}

\item{X23}{A vector of half-panel cross-sectional data 23}

\item{X24}{A vector of half-panel cross-sectional data 24}

\item{X31}{A vector of third-panel cross-sectional data 31}

\item{X32}{A vector of third-panel cross-sectional data 32}

\item{X33}{A vector of third-panel cross-sectional data 33}

\item{X34}{A vector of third-panel cross-sectional data 34}

\item{X35}{A vector of third-panel cross-sectional data 35}

\item{X36}{A vector of third-panel cross-sectional data 36}

\item{X37}{A vector of third-panel cross-sectional data 37}

\item{X38}{A vector of third-panel cross-sectional data 38}

\item{X39}{A vector of third-panel cross-sectional data 39}

\item{h}{A scalar of bandwidth}
}
\value{
A vector of kernel density estimates
}
\description{
Compute TOJ kernel density estimate for T equivalent to 1 modulo 6
}
