% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitytoj.R
\name{tojkdest4}
\alias{tojkdest4}
\title{computing TOJ kernel density estimate for T equivalent to 4 modulo 6}
\usage{
tojkdest4(x, X, X21, X22, X31, X32, X33, X34, X35, X36, X37, X38, X39, h)
}
\arguments{
\item{x}{point at which the density is estimated}

\item{X}{vector of original cross-sectional data}

\item{X21}{vector of half-panel cross-sectional data based on time series 1 ~ T/2}

\item{X22}{vector of half-panel cross-sectional data based on time series (T/2 + 1) ~ T}

\item{X31}{vector of one-third-panel cross-sectional data based on time series 1 ~ floor(T/3)}

\item{X32}{vector of one-third-panel cross-sectional data based on time series (floor(T/3) + 1) ~ (2 * floor(T/3))}

\item{X33}{vector of one-third-panel cross-sectional data based on time series (2 * floor(T/3) + 1) ~ T}

\item{X34}{vector of one-third-panel cross-sectional data based on time series 1 ~ floor(T/3)}

\item{X35}{vector of one-third-panel cross-sectional data based on time series (floor(T/3) + 1) ~ (2 * floor(T/3) + 1)}

\item{X36}{vector of one-third-panel cross-sectional data based on time series (2 * floor(T/3) + 2) ~ T}

\item{X37}{vector of one-third-panel cross-sectional data based on time series 1 ~ ceiling(T/3)}

\item{X38}{vector of one-third-panel cross-sectional data based on time series (ceiling(T/3) + 1) ~ (2 * floor(T/3) + 1)}

\item{X39}{vector of one-third-panel cross-sectional data based on time series (2 * floor(T/3) + 2) ~ T}

\item{h}{bandwidth}
}
\description{
computing TOJ kernel density estimate for T equivalent to 4 modulo 6
}
