% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenttoj.R
\name{tojmomentest0}
\alias{tojmomentest0}
\title{computing TOJ bias-corrected estimate for T equivalent to 0 modulo 6}
\usage{
tojmomentest0(quantity, indices)
}
\arguments{
\item{quantity}{N * (3 * 6) matrix of the means, autocovariances, autocorrelations}

\item{indices}{indices for bootstrap replications}
}
\description{
computing TOJ bias-corrected estimate for T equivalent to 0 modulo 6
}
