% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenttoj.R
\name{tojmomentest1}
\alias{tojmomentest1}
\title{computing TOJ bias-corrected estimate for T equivalent to 1 modulo 6}
\usage{
tojmomentest1(quantity, indices)
}
\arguments{
\item{quantity}{N * (3 * 14) matrix of the means, autocovariances, autocorrelations}

\item{indices}{indices for bootstrap replications}
}
\description{
computing TOJ bias-corrected estimate for T equivalent to 1 modulo 6
}
