% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenttoj.R
\name{tojmomentest2}
\alias{tojmomentest2}
\title{computing TOJ bias-corrected estimate for T equivalent to 2 modulo 6}
\usage{
tojmomentest2(quantity, indices)
}
\arguments{
\item{quantity}{N * (3 * 12) matrix of the means, autocovariances, autocorrelations}

\item{indices}{indices for bootstrap replications}
}
\description{
computing TOJ bias-corrected estimate for T equivalent to 2 modulo 6
}
