% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenttoj.R
\name{tojmomentest3}
\alias{tojmomentest3}
\title{computing TOJ bias-corrected estimate for T equivalent to 3 modulo 6}
\usage{
tojmomentest3(quantity, indices)
}
\arguments{
\item{quantity}{N * (3 * 8) matrix of the means, autocovariances, autocorrelations}

\item{indices}{indices for bootstrap replications}
}
\description{
computing TOJ bias-corrected estimate for T equivalent to 3 modulo 6
}
