% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenttoj.R
\name{tojmomentest4}
\alias{tojmomentest4}
\title{Compute TOJ bias-corrected estimates for T equivalent to 4 modulo 6}
\usage{
tojmomentest4(quantity, indices)
}
\arguments{
\item{quantity}{An N times (3 times 12) matrix of the estimated quantities}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute TOJ bias-corrected estimates for T equivalent to 4 modulo 6
}
