% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{acor}
\alias{acor}
\title{Compute the autocorrelation of time series}
\usage{
acor(y, acor_order)
}
\arguments{
\item{y}{A vector of time series}

\item{acor_order}{A positive integer of the order of autocorrelation}
}
\value{
A scalar of the autocorrelation
}
\description{
Compute the autocorrelation of time series
}
