% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{acov}
\alias{acov}
\title{Compute the autocovariance of time series}
\usage{
acov(y, acov_order)
}
\arguments{
\item{y}{A vector of time series}

\item{acov_order}{A non-negative integer of the order of autocovariance}
}
\value{
A scalar of the autocovariance
}
\description{
Compute the autocovariance of time series
}
