% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{ecdfest}
\alias{ecdfest}
\title{Compute empirical CDF estimate}
\usage{
ecdfest(x, X)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of cross-sectional data}
}
\description{
Compute empirical CDF estimate
}
