% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error1}
\alias{error1}
\title{Error handling 1}
\usage{
error1(data, acov_order, acor_order, R)
}
\arguments{
\item{data}{A matrix of panel data}

\item{acov_order}{A non-negative integer of the order of autocovariance}

\item{acor_order}{A positive integer of the order of autocorrelation}

\item{R}{A positive integer of the number of bootstrap repetitions}
}
\description{
Error handling 1
}
