% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error2}
\alias{error2}
\title{Error handling 2}
\usage{
error2(data, acov_order, acor_order)
}
\arguments{
\item{data}{A matrix of panel data}

\item{acov_order}{A non-negative integer of the order of autocovariance}

\item{acor_order}{A positive integer of the order of autocorrelation}
}
\description{
Error handling 2
}
