% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error3}
\alias{error3}
\title{Error handling 3}
\usage{
error3(data, acov_order, acor_order, mean_bw, acov_bw, acor_bw)
}
\arguments{
\item{data}{A matrix of panel data}

\item{acov_order}{A non-negative integer of the order of autocovariance}

\item{acor_order}{A positive integer of the order of autocorrelation}

\item{mean_bw}{A scalar of bandwidth used for the estimation of
the denisty of mean}

\item{acov_bw}{A scalar of bandwidth used for the estimation of
the denisty of autocovariance}

\item{acor_bw}{A scalar of bandwidth used for the estimation of
the denisty of autocorrelation}
}
\description{
Error handling 3
}
