% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfhpj.R
\name{hpj2_boot}
\alias{hpj2_boot}
\title{Compute bootstrap HPJ empirical CDF estimate for even T}
\usage{
hpj2_boot(quantity, indices)
}
\arguments{
\item{quantity}{An N * 5 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap HPJ empirical CDF estimate for even T
}
