% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfhpj.R
\name{hpjecdfest2}
\alias{hpjecdfest2}
\title{Compute HPJ bias-corrected empirical CDF estimate for odd T}
\usage{
hpjecdfest2(x, X, X1, X2, X3, X4)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of original cross-sectional data}

\item{X1}{A vector of half-panel cross-sectional data 1}

\item{X2}{A vector of half-panel cross-sectional data 2}

\item{X3}{A vector of half-panel cross-sectional data 3}

\item{X4}{A vector of half-panel cross-sectional data 4}
}
\description{
Compute HPJ bias-corrected empirical CDF estimate for odd T
}
