% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityhpj.R
\name{hpjkdest1}
\alias{hpjkdest1}
\title{Compute HPJ kernel density estimates for even T}
\usage{
hpjkdest1(x, X, X1, X2, h)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of cross-sectional data}

\item{X1}{A vector of half-panel cross-sectional data 1}

\item{X2}{A vector of half-panel cross-sectional data 2}

\item{h}{A scalar of bandwidth}
}
\value{
A vector of kenrel density estimates
}
\description{
Compute HPJ kernel density estimates for even T
}
