% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityhpj.R
\name{hpjkdest2}
\alias{hpjkdest2}
\title{Compute HPJ kernel density estimates for odd T}
\usage{
hpjkdest2(x, X, X1, X2, X3, X4, h)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of cross-sectional data}

\item{X1}{A vector of half-panel cross-sectional data 1}

\item{X2}{A vector of half-panel cross-sectional data 2}

\item{X3}{A vector of half-panel cross-sectional data 3}

\item{X4}{A vector of half-panel cross-sectional data 4}

\item{h}{A scalar of bandwidth}
}
\value{
A vector of kernel density estimates
}
\description{
Compute HPJ kernel density estimates for odd T
}
