% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenthpj.R
\name{hpjmomentest1}
\alias{hpjmomentest1}
\title{Compute HPJ bias-corrected estimates of the moments for even T}
\usage{
hpjmomentest1(quantity, indices)
}
\arguments{
\item{quantity}{An N * 3 matrix of the estimated quantities}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute HPJ bias-corrected estimates of the moments for even T
}
