% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{kdest}
\alias{kdest}
\title{Compute kernel density estimates}
\usage{
kdest(x, X, h)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of cross-sectional data}

\item{h}{A scalar of bandwidth}
}
\value{
A vector of kernel density estimates
}
\description{
Compute kernel density estimates
}
