% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{momentest}
\alias{momentest}
\title{Compute the estimates of the moments}
\usage{
momentest(quantity, indices)
}
\arguments{
\item{quantity}{An N * 3 matrix of the estimated quantities}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute the estimates of the moments
}
