% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{ne_boot}
\alias{ne_boot}
\title{Compute bootstrap empirical CDF estimate}
\usage{
ne_boot(quantity, indices)
}
\arguments{
\item{quantity}{An N * 1 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap empirical CDF estimate
}
