% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{toj0_boot}
\alias{toj0_boot}
\title{Compute bootstrap TOJ empirical CDF estimate for T equivalent to 0 modulo 6}
\usage{
toj0_boot(quantity, indices)
}
\arguments{
\item{quantity}{An N * 6 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap TOJ empirical CDF estimate for T equivalent to 0 modulo 6
}
