% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{toj1_boot}
\alias{toj1_boot}
\title{Compute bootstrap TOJ empirical CDF estimate for T equivalent to 1 modulo 6}
\usage{
toj1_boot(quantity, indices)
}
\arguments{
\item{quantity}{An N * 14 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap TOJ empirical CDF estimate for T equivalent to 1 modulo 6
}
