% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{toj2_boot}
\alias{toj2_boot}
\title{Compute bootstrap TOJ empirical CDF estimate for T equivalent to 2 modulo 6}
\usage{
toj2_boot(quantity, indices)
}
\arguments{
\item{quantity}{An N * 12 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap TOJ empirical CDF estimate for T equivalent to 2 modulo 6
}
