% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{toj3_boot}
\alias{toj3_boot}
\title{Compute bootstrap TOJ empirical CDF estimate for T equivalent to 3 modulo 6}
\usage{
toj3_boot(quantity, indices)
}
\arguments{
\item{quantity}{An N * 8 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap TOJ empirical CDF estimate for T equivalent to 3 modulo 6
}
