% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdftoj.R
\name{toj5_boot}
\alias{toj5_boot}
\title{Compute bootstrap TOJ empirical CDF estimate for T equivalent to 5 modulo 6}
\usage{
toj5_boot(quantity, indices)
}
\arguments{
\item{quantity}{N * 14 matrix of estimates}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute bootstrap TOJ empirical CDF estimate for T equivalent to 5 modulo 6
}
