% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitytoj.R
\name{tojkdest0}
\alias{tojkdest0}
\title{Compute TOJ kernel density estimates for T equivalent to 0 modulo 6}
\usage{
tojkdest0(x, X, X21, X22, X31, X32, X33, h)
}
\arguments{
\item{x}{An evaluation point}

\item{X}{A vector of cross-sectional data}

\item{X21}{A vector of half-panel cross-sectional data 21}

\item{X22}{A vector of half-panel cross-sectional data 22}

\item{X31}{A vector of third-panel cross-sectional data 31}

\item{X32}{A vector of third-panel cross-sectional data 32}

\item{X33}{A vector of third-panel cross-sectional data 33}

\item{h}{A scalar of bandwidth}
}
\value{
A vector of kernel density estimates
}
\description{
Compute TOJ kernel density estimates for T equivalent to 0 modulo 6
}
