% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenttoj.R
\name{tojmomentest2}
\alias{tojmomentest2}
\title{Compute TOJ bias-corrected estimate for T equivalent to 2 modulo 6}
\usage{
tojmomentest2(quantity, indices)
}
\arguments{
\item{quantity}{An N times (3 times 12) matrix of the estimated quantities}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute TOJ bias-corrected estimate for T equivalent to 2 modulo 6
}
