% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momenttoj.R
\name{tojmomentest3}
\alias{tojmomentest3}
\title{Compute TOJ bias-corrected estimate for T equivalent to 3 modulo 6}
\usage{
tojmomentest3(quantity, indices)
}
\arguments{
\item{quantity}{An N times (3 times 8) matrix of the estimated quantities}

\item{indices}{A vector of indices for bootstrap repetitions}
}
\description{
Compute TOJ bias-corrected estimate for T equivalent to 3 modulo 6
}
